box/v3: repurpose `catalyst` binary to box entrypoint
