--- cmd/catalyst/conf/catalyst.local.json
@@ -1,184 +1,0 @@
-{
-  "autopushes": null,
-  "bandwidth": {
-    "exceptions": [],
-    "limit": 0
-  },
-  "config": {
-    "accesslog": "LOG",
-    "controller": {
-      "interface": null,
-      "port": null,
-      "username": null
-    },
-    "debug": null,
-    "defaultStream": null,
-    "limits": null,
-    "location": {
-      "lat": 0,
-      "lon": 0,
-      "name": ""
-    },
-    "prometheus": "",
-    "protocols": [
-      {
-        "connector": "AAC"
-      },
-      {
-        "connector": "CMAF"
-      },
-      {
-        "connector": "DTSC"
-      },
-      {
-        "connector": "EBML"
-      },
-      {
-        "connector": "FLV"
-      },
-      {
-        "connector": "H264"
-      },
-      {
-        "connector": "HDS"
-      },
-      {
-        "connector": "HLS"
-      },
-      {
-        "connector": "HTTP"
-      },
-      {
-        "connector": "HTTPTS"
-      },
-      {
-        "connector": "JSON"
-      },
-      {
-        "connector": "MP3"
-      },
-      {
-        "connector": "MP4"
-      },
-      {
-        "connector": "OGG"
-      },
-      {
-        "connector": "RTMP"
-      },
-      {
-        "connector": "RTSP"
-      },
-      {
-        "connector": "SDP"
-      },
-      {
-        "connector": "SRT"
-      },
-      {
-        "connector": "TSSRT"
-      },
-      {
-        "connector": "WAV"
-      },
-      {
-        "connector": "WebRTC"
-      },
-      {
-        "autoAdjustPrice": false,
-        "broadcaster": true,
-        "connector": "livepeer",
-        "currentManifest": false,
-        "datadir": "{{ .DataDir }}/offchain-broadcaster",
-        "httpIngest": true,
-        "initializeRound": false,
-        "j": false,
-        "localVerify": false,
-        "metricsClientIP": false,
-        "metricsPerStream": false,
-        "monitor": true,
-        "orchAddr": "127.0.0.1:8936",
-        "orchestrator": false,
-        "redeemer": false,
-        "reward": false,
-        "rtmpAddr": "127.0.0.1:1936",
-        "testTranscoder": false,
-        "transcoder": false,
-        "version": false
-      },
-      {
-        "autoAdjustPrice": false,
-        "broadcaster": false,
-        "cliAddr": "127.0.0.1:7936",
-        "connector": "livepeer",
-        "currentManifest": false,
-        "datadir": "{{ .DataDir }}/offchain-orchestrator",
-        "httpIngest": false,
-        "initializeRound": false,
-        "j": false,
-        "localVerify": false,
-        "metricsClientIP": false,
-        "metricsPerStream": false,
-        "monitor": true,
-        "orchestrator": true,
-        "redeemer": false,
-        "reward": false,
-        "serviceAddr": "127.0.0.1:8936",
-        "testTranscoder": false,
-        "transcoder": true,
-        "version": false
-      }
-    ],
-    "serverid": null,
-    "sessionInputMode": "14",
-    "sessionOutputMode": "14",
-    "sessionViewerMode": "14",
-    "triggers": null,
-    "trustedproxy": []
-  },
-  "push_settings": {
-    "maxspeed": null,
-    "wait": null
-  },
-  "streams": {
-    "stream": {
-      "name": "stream",
-      "processes": [
-        {
-          "exit_unmask": false,
-          "hardcoded_broadcasters": "[{\"address\":\"http://127.0.0.1:8935\"}]",
-          "process": "Livepeer",
-          "target_profiles": [
-            {
-              "bitrate": 250000,
-              "height": 240,
-              "name": "240p0",
-              "width": 426
-            },
-            {
-              "bitrate": 800000,
-              "height": 360,
-              "name": "360p0",
-              "width": 640
-            },
-            {
-              "bitrate": 1600000,
-              "height": 480,
-              "name": "480p0",
-              "width": 854
-            },
-            {
-              "bitrate": 3000000,
-              "height": 720,
-              "name": "720p0",
-              "width": 1280
-            }
-          ]
-        }
-      ],
-      "source": "push://",
-      "stop_sessions": false
-    }
-  },
-  "ui_settings": null
-}
